+++ attributes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: hana_client
+# Cookbook Name:: hana-client
 #
 # Copyright 2016, SAP
 #
@@ -26,13 +26,12 @@ default['sap']['sapcar'] = nil
 # The complete URL to the SAP SAR package for the hana client to be installed
 default['sap']['hanaclient'] = nil
 
-
 #########################
 ## Advanced Attributes ##
 #########################
 
 # This is where the hana client will live on your system.
-default['hana_client']['root_install_folder'] = "c:\\sap"
+default['hana-client']['root_install_folder'] = 'c:\sap'
 
 # Signals the removal of any existing hana clients in the install folder.
-default['hana_client']['uninstall_reinstall'] = false
+default['hana-client']['uninstall_reinstall'] = false