+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'hana_client'
+name             'hana-client'
 maintainer       'SAP'
 maintainer_email 'Dan-Joe.Lopez@sap.com'
 license          'Apache 2.0'
-description      'Installs/Configures Hana Client for Windows.'
+description      'Installs/Configures Hana Client for Windows only'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '1.0.0'