+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: hana_client
+# Cookbook Name:: hana-client
 # Recipe:: default
 #
 # Copyright 2016, SAP
@@ -17,32 +17,37 @@
 # limitations under the License.
 #
 
+# Some vars for readability
+destination = node['hana-client']['root_install_folder']
+extract_to_dir = Chef::Config[:file_cache_path] + '\sap_temp'
+local_installer = extract_to_dir + '\SAP_HANA_CLIENT\hdbinst.exe'
+
+# Get current state
+pkg_extracted = ::Dir.exist?(extract_to_dir)
+installed = ::File.exist?("#{destination}\\hdbclient\\install\\hdbuninst.exe")
+
 # If requested, uninistall previous HANA client versions.
-hana_client "#{node['hana_client']['root_install_folder']}" do
+hana_client node['hana-client']['root_install_folder'] do
   action :uninstall
-  only_if { node['hana_client']['uninstall_reinstall'] }
+  only_if { node['hana-client']['uninstall_reinstall'] }
 end
 
 # Extract the SAR file
-remote_source = node['sap']['hanaclient']
-sar_name = remote_source.split('/')[-1]
-temp_path = "c:"
-
-hana_client_sap_media "#{temp_path}" do
-  remote_path remote_source
+hana_client_sap_media extract_to_dir do
+  remote_path node['sap']['hanaclient']
   sapcar node['sap']['sapcar']
   action :extract
-  not_if  { ::Dir.exist?("#{temp_path}\\SAP_HANA_CLIENT") || ::File.exist?("#{node['hana_client']['root_install_folder']}\\hdbclient\\install\\hdbuninst.exe") }
+  not_if { pkg_extracted || installed }
 end
 
 # Install the client from the extracted SAR
-hana_client "#{node['hana_client']['root_install_folder']}" do
-  installer "#{temp_path}\\SAP_HANA_CLIENT\\hdbinst.exe"
+hana_client destination do
+  installer local_installer
   action :install
 end
 
 # Cleanup the installer source file.
-directory "#{temp_path}\\SAP_HANA_CLIENT\\" do
+directory extract_to_dir do
   recursive true
   action :delete
 end